% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reviews_over.R
\name{get_reviews_over}
\alias{get_reviews_over}
\title{Get reviews over a given mark}
\usage{
get_reviews_over(points)
}
\arguments{
\item{points}{Lower bound mark}
}
\value{
A `data.frame`
}
\description{
Returns a data frame with information related
to appartments with a mark above a given number
}
\examples{
get_reviews_over(5)
}
