% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input.R
\name{inputReviewUI}
\alias{inputReviewUI}
\alias{inputReviewServer}
\title{Input Window Module}
\usage{
inputReviewUI(id)

inputReviewServer(id, path_to_data)
}
\arguments{
\item{id}{Shiny module.}

\item{path_to_data}{Path to rds.}
}
\value{
A UI module

A Server module.
}
\description{
Input Window Module
}
\section{Functions}{
\itemize{
\item \code{inputReviewServer()}: Server side of the module.

}}
