% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_show_map.R
\name{showMapUI}
\alias{showMapUI}
\alias{showMapServer}
\title{Show Map}
\usage{
showMapUI(id)

showMapServer(id, df_locs)
}
\arguments{
\item{id}{Shiny module}

\item{df_locs}{Reactive data frame with locs info}
}
\value{
A leaflet shiny output

A server shiny module.
}
\description{
Shiny module functions for displaying leaflet map
}
\section{Functions}{
\itemize{
\item \code{showMapServer()}: Server module

}}
