% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_show_marker.R
\name{showMarkerUI}
\alias{showMarkerUI}
\alias{showMarkerServer}
\alias{create_html_marker_info}
\title{Display marker info}
\usage{
showMarkerUI(id)

showMarkerServer(id, df_locs)

create_html_marker_info(date, address, flat, mark, comment)
}
\arguments{
\item{id}{Shiny module}

\item{df_locs}{Reactive data frame with locs info}

\item{date}{Review date.}

\item{address}{Address from marker.}

\item{flat}{Flat from marker.}

\item{mark}{Mark from marker.}

\item{comment}{Comment review from marker.}
}
\value{
A Shiny Module UI.

A shiny module server

A `htmltools::HTML` object
}
\description{
Display marker information.
}
\section{Functions}{
\itemize{
\item \code{showMarkerServer()}: Server side for displaying marker info.

\item \code{create_html_marker_info()}: Create HTML for marker info display.

}}
