% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_with_new_address.R
\name{update_with_new_address}
\alias{update_with_new_address}
\title{Update RDS file with provided address}
\usage{
update_with_new_address(google_key, path, new_address, flat, mark, comment)
}
\arguments{
\item{google_key}{Google App key}

\item{path}{Path to RDS file}

\item{new_address}{Address whose coordinates you want to query.}

\item{flat}{Text with flat.}

\item{mark}{Number from 1 to 10 with the review's mark.}

\item{comment}{Text with the review's comment.}
}
\value{
Nothing
}
\description{
Updates the RDS file with locations with the location of the new address
}
